% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/inside_functions.r
\name{check_path}
\alias{check_path}
\title{Function check excistence of folder and file}
\usage{
check_path(path, filename, defaultName)
}
\description{
Function check excistence of folder and file
}

