% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/xcms_proceed.r
\name{combine_ms_data_rt}
\alias{combine_ms_data_rt}
\title{Combine data from xcms results for the plotting purpose}
\usage{
combine_ms_data_rt(pos, neg)
}
\arguments{
\item{pos}{a data.frame from xcms \link[xcms]{diffreport}. Positive m/z
values}

\item{neg}{a data.frame from xcms \link[xcms]{diffreport}. Negative m/z
values}
}
\value{
A list with extracted m/z, folds, tstat and pvalues.
\item{mz}{a vector of all m/z. First positive, then negative.}
\item{folds}{a vector, showing how the mean value of \code{class2} differs
  from the mean value in \code{class1}. When the mean of \code{class1} equals
  to 0 or NA \code{fold} value become \code{Inf}. But that means that no such
  peak is present in negative control, so fold value is changed to maximum
  possible in this dataset}
\item{tstat}{a vector with Welch's t test statistic of fold values. Can be
negative if \code{class1} is greater then \code{class2}}
\item{pvalue}{a vector of p-value of t-statistic}
}
\description{
Take results from xcms \link[xcms]{diffreport} for both positive and negative
modes, extract all given m/z and corresponding statistical values.
}
\details{
Returned list is a combination of \code{pos} and \code{neg} (in this
  order) values.

  Sometimes \code{folds} can be \code{Inf}, so they are converted to
  \code{max} value.
\preformatted{xcmsSet.R:diffreport()
...
## c1 and c2 are columns of class1 and class2 resp.
c1 <- which(classlabel %in% class1)
c2 <- which(classlabel %in% class2)

## calculate both class means
mean1 <- rowMeans(values[,c1], na.rm = TRUE)
mean2 <- rowMeans(values[,c2], na.rm = TRUE)

## Calculate fold change.
## For foldchange <1 set fold to 1/fold
fold <- mean2 / mean1
fold[!is.na(fold) & fold < 1] <- 1/fold[!is.na(fold) & fold < 1]
}
}
\seealso{
\link[xcms]{diffreport}
}

