\name{gerbiochem-package}
\alias{gerbiochem-package}
\alias{gerbiochem}
\docType{package}
\title{
\packageTitle{gerbiochem}
}
\description{
\packageDescription{gerbiochem}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{gerbiochem}
\packageIndices{gerbiochem}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{gerbiochem}

Maintainer: \packageMaintainer{gerbiochem}
}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
