% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/xcms_proceed.r
\name{get_xcms_result}
\alias{get_xcms_result}
\title{Default scripting for getting xcms data}
\usage{
get_xcms_result(directory, retMethod = "loess", ...)
}
\arguments{
\item{directory}{a location of grouped chromatograms in accesable format}

\item{retMethod}{What method should be used in \link[xcms]{retcor} function.
Default method is "loess." Another possible variant is "obiwarp"}

\item{...}{parameters passed to \link[xcms]{diffreport} method}
}
\description{
Function contains allmost all needable commands to get results from xcms
package. Still the main command (\link[xcms]{diffreport}) is required as an
argument
}
\details{
xcms \link[xcms]{diffreport} method returns the data.frame in which
  retention time is showed in seconds. \code{\link{get_xcms_result}}
  transforms them into minute values
}

