% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_data.r
\name{plot_activity}
\alias{plot_activity}
\title{Plot chromatogram with radiactivity as a signal}
\usage{
plot_activity(activity, method, toPDF = FALSE, filename = "", path = "/")
}
\arguments{
\item{activity}{the vector with data to plot}

\item{method}{a data.frame with chromatographical method used to produce data}

\item{toPDF}{an optional logical value. Default is \code{FALSE}. Should the
plot be rendered in pdf device.}

\item{filename}{a name of the file without extention (.pdf)}

\item{path}{an optional string. Path to the folder, where file should be
  created. If not specified, file created in working directory.}
}
\description{
This function shows only bars that was higher then background. Drawing line
showed all fractions
}
\details{
If \code{path} directs to unexistent folder, a new folder will be
  created under this path
}

