% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_data.r
\name{plot_mz}
\alias{plot_mz}
\title{Generic plot function for MS-spectrum}
\usage{
plot_mz(mz, folds, RT = 0, tstat = FALSE, pvalue = FALSE)
}
\arguments{
\item{mz}{a numerical vector. Contains negative and positive m/z together}

\item{folds}{a numerical vector. Each number related to the number in
\code{mz} parameter. The height of the line.}

\item{RT}{optional number. Used as a title}

\item{tstat}{an optional numerical vector. Each number related to the number
in \code{mz} parameter. Value represents Welch's t test. See \emph{details}}

\item{pvalue}{pvalue of t-statistic}
}
\description{
Plot m/z grouped together by specific time. The significance of peaks is
based on the difference between treatment and negative control
(\code{\strong{folds}}) (or if set otherwise in other functions). If other
statistical values are passed, they will be printed next to the peaks
}
\details{
Description of \code{\strong{tstat}} in \link[xcms]{diffreport}
  function (xcms package): Welch's two sample t-statistic, positive for
  analytes having greater intensity in \code{class2}, negative for analytes having
  greater intensity in \code{class1}
}
\seealso{
\code{\link{plot_choosed_rt}}
}

