% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_data.r
\name{plot_TEIC}
\alias{plot_TEIC}
\title{Display Total Ion Chromatogram}
\usage{
plot_TEIC(pos, neg, datanames, rtRange = FALSE, title = "", toPDF = FALSE,
  filename = "")
}
\arguments{
\item{pos}{a list of data from xcms-package, containing only positive m/z
values}

\item{neg}{a list of data from xcms-package, containing only negative m/z
values}

\item{datanames}{a list with three string values, related to the names of
treatment in the datasets. See \emph{details} for the list structure}

\item{rtRange}{an optional numerical vector with a time window of data to
display. If \code{FALSE} (the default) the range function will be used to
determine window to include all values}

\item{title}{(optional) an overall title of the plot}

\item{toPDF}{logical (optional, default is \code{FALSE}). Should the plot be
rendered in pdf device.}

\item{filename}{path string. Should specify if \code{toPDF} parameter is
\code{TRUE}. Should have an extension .pdf. Does not create a folder}
}
\description{
Display two/three groups in one plot, where y-axis is a time, and x-axis is a
m/z (positive and negative).
}
\details{
\code{\strong{datanames}} Package xcms generate a dataset with
  number of peaks in each time point. The number and names of the columns in
  the dataset depends on your initial file structure. To get access to these
  columns \code{\link{plot_TEIC}} need the names of the folder. Usually there
  are two or three groups. This function use three values - \emph{treatment,
  neg_control, pos_control} in this order. So the structure of datanames
  should be \code{datanames$treatment, datanames$neg_control,
  datanames$pos_control} If there are only two groups, function will use only
  \emph{treatment} and \emph{neg_control} parameters

  \code{\strong{filename}} If using \code{\link{plot_TEIC}} directly, check
  directory first with \code{\link[base]{dir.exists}} and create one after
  with \code{\link[base]{dir.create}}
}

