% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/output.r
\name{results_to_excel}
\alias{results_to_excel}
\title{Print all given data into excel table}
\usage{
results_to_excel(data, filename = "", path = "/")
}
\arguments{
\item{data}{a list with data.frames to save. Each element in the list is one
data.frame. Name of the element will be a sheet name.}

\item{filename}{a name of the file without extention (neither .xls or .xlsx)}

\item{path}{an optional string. Path to the folder, where file should be
  created. If not specified, file created in working directory.}
}
\description{
For all given data.frames create a sheet in the excel table and save them
there
}
\details{
If \code{path} directs to unexistent folder, a new folder will be
  created under this path
}

